requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Machine Learning
scikit-learn>=1.4.0
numpy>=1.26.3
pandas>=2.2.0
joblib>=1.3.2

# Feature Engineering
scipy>=1.12.0

# Audio Processing (for future audio upload feature)
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
